Keep user.entity types in one file
so FE can use it in build.
